!Rewrite user session create logic
